requirements.txt
streamlit>=1.41.0
langchain>=0.3.18
langchain-community>=0.3.19
langchain-core>=0.3.40
chromadb>=0.6.2
sentence-transformers>=3.4.1
ollama>=0.4.7
pypdf>=5.3.1
pytesseract>=0.3.13
Pillow>=10.4.0
pydantic>=2.10.6
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
langchain-chroma>=0.2.0
langchain-huggingface>=0.1.2
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.6
six>=1.16.0
tiktoken>=0.8.0
unstructured>=0.11.8
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pydub>=0.25.1
audio-recorder-streamlit>=0.0.9
numpy>=1.24.0
duckduckgo-search>=5.0.0
imageio-ffmpeg>=0.4.9
